[Submit] Share boilerplate between submit commands
